// contracts/generated/typescript/models/RevisionStorageStats.ts
export interface RevisionStorageStats {
    /**
     * API 与 Worker 是否已装配 Revision 对象存储
     * @type {boolean}
     * @memberof RevisionStorageStats
     */
    archiveAvailable: boolean;
    /**
     *
     * @type {number}
     * @memberof RevisionStorageStats
     */
    retentionSeconds: number;
    /**
     *
     * @type {number}
     * @memberof RevisionStorageStats
     */
    defaultBatchSize: number;
    /**
     *
     * @type {number}
     * @memberof RevisionStorageStats
     */
    hotSnapshots: number;
    /**
     *
     * @type {number}
     * @memberof RevisionStorageStats
     */
    coldSnapshots: number;
    /**
     *
     * @type {number}
     * @memberof RevisionStorageStats
     */
    hotBytes: number;
    /**
     *
     * @type {number}
     * @memberof RevisionStorageStats
     */
    coldBytes: number;
    /**
     *
     * @type {number}
     * @memberof RevisionStorageStats
     */
    eligibleSnapshots: number;
    /**
     *
     * @type {Date}
     * @memberof RevisionStorageStats
     */
    oldestHotCreatedAt?: Date;
}